Moved FileChecksumUUIDAccessionGenerator to package file
